requirements.txt
# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
alembic==1.13.1

# ETL and Data Processing
pandas==2.1.4
numpy==1.26.3
requests==2.31.0
python-dotenv==1.0.0

# API and Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
pytz==2023.3

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0

# Airflow (separate installation recommended)
# apache-airflow==2.8.0

# Development tools
ipython==8.19.0
black==23.12.1
flake8==7.0.0

# Logging and monitoring
loguru==0.7.2


